feat: removed all pages for sponsorship/project/bounty
